+++ tests/test_io.py
@@ -126,7 +126,7 @@ def test_loadmol_sdf(molfile, natoms: int) -> None:
 
 @pytest.mark.parametrize(
     "molfile, natoms",
-    [("1cbr_ligand.mol2", 49)],
+    [("benzene.mol2", 12), ("1cbr_ligand.mol2", 49)],
 )
 def test_loadmol_mol2(molfile, natoms: int) -> None:
 
@@ -140,6 +140,31 @@ def test_loadmol_mol2(molfile, natoms: int) -> None:
     [("1cbr_docking.sdf", 22)],
 )
 def test_loadallmols_sdf(molfile, natoms: int) -> None:
+
+    ms = io.loadallmols(os.path.join(molpath, molfile))
+
+    assert len(ms) == 10
+
+    for m in ms:
+        assert len(m) == natoms
+
+
+@pytest.mark.parametrize(
+    "molfile, natoms",
+    [("benzene.sdf.gz", 12), ("benzene.mol2.gz", 12), ("1a99_ligand.pdb.gz", 20)],
+)
+def test_loadmol_gz(molfile, natoms: int) -> None:
+
+    m = io.loadmol(os.path.join(molpath, molfile))
+
+    assert len(m) == natoms
+
+
+@pytest.mark.parametrize(
+    "molfile, natoms",
+    [("1cbr_docking.sdf.gz", 22)],
+)
+def test_loadallmols_sdf_gz(molfile, natoms: int) -> None:
 
     ms = io.loadallmols(os.path.join(molpath, molfile))